+++ crates/ntscrs/src/noise.rs
@@ -8,20 +8,20 @@ const VALUE_SEED: u64 = 2;
 /// `jitter` is the amount to move around the points which are interpolated to make value noise.
 /// This is useful for avoiding periodic artifacts.
 /// `seed` is the random seed.
-pub fn sample_noise(t: f64, jitter: f64, seed: u64) -> f64 {
+pub fn sample_noise(t: f32, jitter: f32, seed: u64) -> f32 {
     let left_coord = t as u64;
     let cellspace_coord = t.fract();
 
     let mut left_jitter = (Seeder::new(left_coord)
         .mix(seed)
         .mix(JITTER_SEED)
-        .finalize::<f64>()
+        .finalize::<f32>()
         - 0.5)
         * jitter;
     let mut right_jitter = (Seeder::new(left_coord.wrapping_add(1))
         .mix(seed)
         .mix(JITTER_SEED)
-        .finalize::<f64>()
+        .finalize::<f32>()
         - 0.5)
         * jitter;
 
@@ -30,7 +30,7 @@ pub fn sample_noise(t: f64, jitter: f64, seed: u64) -> f64 {
         left_jitter = (Seeder::new(left_coord.wrapping_sub(1))
             .mix(seed)
             .mix(JITTER_SEED)
-            .finalize::<f64>()
+            .finalize::<f32>()
             - 0.5)
             * jitter;
         (-1.0, left_coord.wrapping_sub(1))
@@ -39,7 +39,7 @@ pub fn sample_noise(t: f64, jitter: f64, seed: u64) -> f64 {
         right_jitter = (Seeder::new(left_coord.wrapping_add(2))
             .mix(seed)
             .mix(JITTER_SEED)
-            .finalize::<f64>()
+            .finalize::<f32>()
             - 0.5)
             * jitter;
         (1.0, left_coord.wrapping_add(1))
@@ -48,8 +48,8 @@ pub fn sample_noise(t: f64, jitter: f64, seed: u64) -> f64 {
     };
     let mut dist =
         (cellspace_coord - (left_jitter + dist_offset)) / (right_jitter + 1.0 - left_jitter);
-    let left_rand: f64 = Seeder::new(rand_coord).mix(seed).mix(VALUE_SEED).finalize();
-    let right_rand: f64 = Seeder::new(rand_coord.wrapping_add(1))
+    let left_rand: f32 = Seeder::new(rand_coord).mix(seed).mix(VALUE_SEED).finalize();
+    let right_rand: f32 = Seeder::new(rand_coord.wrapping_add(1))
         .mix(seed)
         .mix(VALUE_SEED)
         .finalize();
